\name{Wine}
\encoding{latin1}
\docType{data}
\alias{Wine}
\title{Wine}
\description{
 The \code{Wine} data file has both learning and test data.
}
\usage{Wine}
\format{
  This data file contains the following items:
  \describe{
    \item{x.learning}{94 spectra (lines), sampled with 256 points (columns)}
    \item{y.learning}{Corresponding level of alcohol, in a column vector (94 lines).}
    \item{x.test}{30 spectra (lines), sampled with 256 points (columns)}
    \item{y.test}{Corresponding level of alcohol, in a column vector (30 lines).}
  }
}
\details{
To estimate the level of alcohol of a wine from its observed mean infrared spectrum.
}
\source{
This database has been provided by Prof. Marc Meurens, Université catholique de Louvain, BNUT unit.
Please acknowledge the origin of this database if you use it and/or publish about it.
It can be found at \url{http://www.ucl.ac.be/mlg/index.php?page=DataBases}
  }
\keyword{datasets}


